microblog/app/requirements.txt
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Assets==2.0

Flask-Login==0.5.0
Flask-WTF==0.14.3
WTForms==2.3.1
email-validator==1.1.0

Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.16
flask-migrate
PyMySQL==0.9.3

lessc==0.1.3
lesscpy==0.14.0
nodeenv==1.3.5
python-dotenv==0.13.0
cssmin==0.2.0

flask-httpauth

